Milestone-4_Multi-Model AI Agent for Automated Health Diagnostics/requirements.txt
pandas
opencv-python
pytesseract
pillow
streamlit
PyMuPDF
pdf2image
reportlab